Disregard context if it's any falsy value
Currently when generating the key for an entry we only disregard if
context is NULL. Unfortunately in some recent cases context has been an
empty string and the key calculated included chr(4) in the beginning.

This lead to some wring comparisons and ultimately including the same
oriignal multiple times during import.