GlotPress as a WordPress plugin, first pass.
This initial pass switches from using the BackPress external to a WordPress install via `require 'wp-load.php';`. This took about 2 hours. With some additional work, it can be pivoted fully to a plugin. This would result in the deletion of additional code (more than has already been deleted here), given that WordPress can handle authentication, users, etc., most initialization that's currently happening in both gp-settings.php and wp-load.php, and such.

I don't see a need to abandon GlotPress's routes, things, or templates in any kind of switch to WP_Rewrite, custom post types, or a theme. That it "just works" by swapping out some files is pretty killer in and of itself. It will make it easier for people to contribute, work with it, and such.

There are surely some little issues that will be caught in a second pass. As I said, this was done in just two hours (and yes, I got it fully working locally in that time). As an example todo: we need to merge some POMO changes to upstream WordPress; some hook names like `plugins_loaded` also clash.

As for performance: locally, I was seeing 3 MB (GlotPress standalone) vs 6 MB (GlotPress with WordPress). That's peanuts and is well worth bringing GlotPress into the comfort zone that is WordPress. Additionally, this will provide significant benefits for feature development and general integration between GlotPress and WordPress.